requirements.txt
pulp==2.8.0
matplotlib==3.8.4
gurobipy==12.0.1
pandas==2.2.2
scipy==1.14.1
networkx==3.3
simpy==4.1.1
lxml==5.3.0
pyyaml==6.0.2
seaborn==0.13.2
psutil==5.9.0